Refactor ApiClient and Session classes
Merge ApiClient and ApiSession.
Move ApiKey to HttpClient defaultheaders.
TODO: rename all ApiSession refs to ApiClient.
